Remove unused JSONLog marshaling
